+++ Cargo.toml
@@ -2,17 +2,15 @@
 name = "apalis-redis"
 version = "1.0.0-alpha.1"
 authors = ["Njuguna Mureithi <mureithinjuguna@gmail.com>"]
-edition.workspace = true
-repository.workspace = true
-readme = "../../README.md"
-
+edition = "2024"
+repository = "https://github.com/apalis-dev/apalis-redis"
 license = "MIT"
-description = "Redis Storage for apalis: use Redis for background jobs and message queueing"
+description = "Background task processing for rust using apalis and redis"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-apalis-core = { path = "../../packages/apalis-core", version = "1.0.0-alpha.1", default-features = false, features = [
+apalis-core = { version = "1.0.0-alpha.1", default-features = false, features = [
     "sleep",
     "json",
 ] }
@@ -33,12 +31,13 @@ async-std = { version = "1.13.0", optional = true }
 thiserror = "2.0.0"
 event-listener = "5.4.0"
 tower = { version = "0.5" }
+ulid = "1.2.1"
 
 
 [dev-dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 # email-service = { path = "../../examples/email-service" }
-apalis-core = { path = "../apalis-core", features = ["test-utils"] }
+apalis-core = { version = "1.0.0-alpha.1", features = ["test-utils"] }
 bincode = "2.0.1"
 
 [features]
