+++ lua/register_consumer.lua
